♻️ remove UIController wrapper methods delegating to page builders
Routes now call library_pages, settings_pages, and sonos_pages directly.
Wrapper-focused tests moved to tests/…/ui_pages/test_library_page_builder.py.
