+++ jukebox/adapters/inbound/admin/ui_controller.py
@@ -1,4 +1,3 @@
-from collections.abc import AsyncIterator
 from typing import Annotated
 
 from jukebox.shared.dependency_messages import optional_extra_dependency_message
@@ -18,24 +17,20 @@ except ModuleNotFoundError as e:
 from pydantic import BaseModel, Field
 
 from jukebox.adapters.inbound.admin.api_controller import APIController
-from jukebox.adapters.inbound.admin.ui_pages.library import DiscForm, DiscTable, LibraryUIPageBuilder
+from jukebox.adapters.inbound.admin.ui_pages.library import DiscForm, LibraryUIPageBuilder
 from jukebox.adapters.inbound.admin.ui_pages.settings import SettingsUIPageBuilder
 from jukebox.adapters.inbound.admin.ui_pages.sonos import SonosSelectionForm, SonosUIPageBuilder
-from jukebox.domain.entities import CurrentTagStatus, Disc, DiscMetadata, DiscOption
+from jukebox.domain.entities import Disc, DiscMetadata, DiscOption
 from jukebox.domain.use_cases.library.add_disc import AddDisc
 from jukebox.domain.use_cases.library.edit_disc import EditDisc
 from jukebox.domain.use_cases.library.get_current_tag_status import GetCurrentTagStatus
 from jukebox.domain.use_cases.library.get_disc import GetDisc
 from jukebox.domain.use_cases.library.list_discs import ListDiscs
 from jukebox.domain.use_cases.library.remove_disc import RemoveDisc
-from jukebox.settings.definitions import (
-    EditableSettingDisplay,
-    get_setting_definition,
-)
+from jukebox.settings.definitions import get_setting_definition
 from jukebox.settings.errors import SettingsError
 from jukebox.settings.selected_sonos_group_repository import SettingsSelectedSonosGroupRepository
 from jukebox.settings.service_protocols import SettingsService
-from jukebox.settings.types import JsonObject, JsonValue
 from jukebox.sonos.discovery import SonosDiscoveryError
 from jukebox.sonos.selection import SaveSonosSelection
 from jukebox.sonos.service import SonosService
@@ -81,12 +76,12 @@ class UIController(APIController):
 
         @self.app.get("/api/ui/", response_model=FastUI, response_model_exclude_none=True)
         def list_discs(toast: str | None = None) -> list[AnyComponent]:
-            return self._build_index_page_components(toast=toast)
+            return self.library_pages.build_index_page_components(toast=toast)
 
         @self.app.get("/api/ui/current-tag-banner/events")
         async def get_current_tag_banner_events(request: Request) -> StreamingResponse:
             return StreamingResponse(
-                self._current_tag_banner_event_stream(request),
+                self.library_pages.current_tag_banner_event_stream(request),
                 media_type="text/event-stream",
                 headers={
                     "Cache-Control": "no-cache",
@@ -97,9 +92,11 @@ class UIController(APIController):
 
         @self.app.get("/api/ui/discs/new", response_model=FastUI, response_model_exclude_none=True)
         def new_disc_form(prefill: str | None = None) -> list[AnyComponent]:
-            return self._build_form_page_components(
+            return self.library_pages.build_form_page_components(
                 title="Add disc",
-                form_components=self._build_new_disc_form_components(prefill_current=(prefill == "current")),
+                form_components=self.library_pages.build_new_disc_form_components(
+                    prefill_current=(prefill == "current")
+                ),
             )
 
         @self.app.post("/api/ui/discs", response_model=FastUI, response_model_exclude_none=True)
@@ -124,9 +121,9 @@ class UIController(APIController):
 
         @self.app.get("/api/ui/discs/{tag_id}/edit", response_model=FastUI, response_model_exclude_none=True)
         def edit_disc_form(tag_id: str) -> list[AnyComponent]:
-            return self._build_form_page_components(
+            return self.library_pages.build_form_page_components(
                 title=f"Edit disc {tag_id}",
-                form_components=self._build_edit_disc_form_components(tag_id),
+                form_components=self.library_pages.build_edit_disc_form_components(tag_id),
             )
 
         @self.app.post("/api/ui/discs/{tag_id}", response_model=FastUI, response_model_exclude_none=True)
@@ -166,9 +163,9 @@ class UIController(APIController):
 
         @self.app.get("/api/ui/discs/{tag_id}/delete", response_model=FastUI, response_model_exclude_none=True)
         def delete_disc_confirmation(tag_id: str) -> list[AnyComponent]:
-            return self._build_form_page_components(
+            return self.library_pages.build_form_page_components(
                 title=f"Delete disc {tag_id}",
-                form_components=self._build_delete_disc_form_components(tag_id),
+                form_components=self.library_pages.build_delete_disc_form_components(tag_id),
             )
 
         # Fast-UI buttons and forms do not support the DELETE method directly. So we cannot call DELETE on
@@ -186,11 +183,11 @@ class UIController(APIController):
 
         @self.app.get("/api/ui/settings", response_model=FastUI, response_model_exclude_none=True)
         def settings_page(toast: str | None = None, toast_message: str | None = None) -> list[AnyComponent]:
-            return self._build_settings_page_components(toast=toast, toast_message=toast_message)
+            return self.settings_pages.build_settings_page_components(toast=toast, toast_message=toast_message)
 
         @self.app.get("/api/ui/settings/{setting_path}/edit", response_model=FastUI, response_model_exclude_none=True)
         def edit_setting_form(setting_path: str) -> list[AnyComponent]:
-            return self._build_settings_edit_page_components(setting_path)
+            return self.settings_pages.build_settings_edit_page_components(setting_path)
 
         @self.app.post("/api/ui/settings/{setting_path}", response_model=FastUI, response_model_exclude_none=True)
         async def update_setting(
@@ -202,13 +199,15 @@ class UIController(APIController):
                 raise HTTPException(status_code=404, detail=f"Unknown setting path: {setting_path}")
 
             try:
-                patch = self._build_settings_patch(setting_path, form.value)
+                patch = self.settings_pages.build_settings_patch(setting_path, form.value)
                 result = self.settings_service.patch_persisted_settings(patch)
             except ValueError as err:
                 raise self._field_validation_error("value", str(err))
             except SettingsError as err:
-                if self._persisted_value_matches(setting_path, self._lookup_optional_dotted_path(patch, setting_path)):
-                    return self._build_settings_success_response(
+                if self.settings_pages.persisted_value_matches(
+                    setting_path, self.settings_pages.lookup_optional_dotted_path(patch, setting_path)
+                ):
+                    return self.settings_pages.build_settings_success_response(
                         "Settings saved, but effective settings are still unavailable."
                     )
                 raise self._field_validation_error("value", str(err))
@@ -217,15 +216,15 @@ class UIController(APIController):
             except Exception as err:
                 raise HTTPException(status_code=500, detail=f"Server error: {str(err)}")
 
-            return self._build_settings_success_response(str(result["message"]))
+            return self.settings_pages.build_settings_success_response(str(result["message"]))
 
         @self.app.post("/api/ui/settings/{setting_path}/reset", response_model=FastUI, response_model_exclude_none=True)
         async def reset_setting(setting_path: str) -> list[AnyComponent]:
-            return self._reset_setting(setting_path)
+            return self.settings_pages.reset_setting(setting_path)
 
         @self.app.get("/api/ui/sonos", response_model=FastUI, response_model_exclude_none=True)
         def sonos_page(toast: str | None = None, toast_message: str | None = None) -> list[AnyComponent]:
-            return self._build_sonos_page_components(toast=toast, toast_message=toast_message)
+            return self.sonos_pages.build_sonos_page_components(toast=toast, toast_message=toast_message)
 
         @self.app.get("/api/ui/sonos/edit", response_model=FastUI, response_model_exclude_none=True)
         def edit_sonos_form(
@@ -236,7 +235,7 @@ class UIController(APIController):
             field_errors = None
             if error_message:
                 field_errors = {self._sonos_field_name_for_error(error_message): error_message}
-            return self._build_sonos_edit_page_components(
+            return self.sonos_pages.build_sonos_edit_page_components(
                 error_message=error_message,
                 field_errors=field_errors,
                 submitted_uids=uids,
@@ -292,38 +291,6 @@ class UIController(APIController):
             c.FireEvent(event=GoToEvent(url=f"/?toast={toast_event_name}")),
         ]
 
-    def _build_settings_success_response(self, message: str) -> list[AnyComponent]:
-        return self.settings_pages.build_settings_success_response(message)
-
-    def _reset_setting(self, setting_path: str) -> list[AnyComponent]:
-        return self.settings_pages.reset_setting(setting_path)
-
-    def _build_sonos_page_components(
-        self,
-        toast: str | None = None,
-        toast_message: str | None = None,
-        error_message: str | None = None,
-    ) -> list[AnyComponent]:
-        return self.sonos_pages.build_sonos_page_components(
-            toast=toast,
-            toast_message=toast_message,
-            error_message=error_message,
-        )
-
-    def _build_sonos_edit_page_components(
-        self,
-        error_message: str | None = None,
-        field_errors: dict[str, str] | None = None,
-        submitted_uids: list[str] | None = None,
-        submitted_coordinator_uid: str | None = None,
-    ) -> list[AnyComponent]:
-        return self.sonos_pages.build_sonos_edit_page_components(
-            error_message=error_message,
-            field_errors=field_errors,
-            submitted_uids=submitted_uids,
-            submitted_coordinator_uid=submitted_coordinator_uid,
-        )
-
     def _reset_sonos_selection(self) -> list[AnyComponent]:
         selected_group_repository = SettingsSelectedSonosGroupRepository(self.settings_service)
         try:
@@ -333,7 +300,7 @@ class UIController(APIController):
                 return self.sonos_pages.build_sonos_success_response(
                     "Sonos selection cleared, but effective settings are still unavailable."
                 )
-            return self._build_sonos_page_components(error_message=str(err))
+            return self.sonos_pages.build_sonos_page_components(error_message=str(err))
         except HTTPException:
             raise
         except Exception as err:
@@ -371,116 +338,6 @@ class UIController(APIController):
             return message
 
         return f"{prefix}{speaker.name} [{speaker.uid}]"
-
-    def _build_index_page_components(self, toast: str | None = None) -> list[AnyComponent]:
-        return self.library_pages.build_index_page_components(toast=toast)
-
-    def _build_settings_page_components(
-        self,
-        toast: str | None = None,
-        toast_message: str | None = None,
-    ) -> list[AnyComponent]:
-        return self.settings_pages.build_settings_page_components(toast=toast, toast_message=toast_message)
-
-    def _build_settings_section_components(
-        self,
-        section: str,
-        settings: list[EditableSettingDisplay],
-    ) -> list[AnyComponent]:
-        return self.settings_pages.build_settings_section_components(section, settings)
-
-    def _build_settings_row(self, setting: EditableSettingDisplay, index: int) -> AnyComponent:
-        return self.settings_pages.build_settings_row(setting, index)
-
-    def _build_settings_edit_page_components(
-        self,
-        setting_path: str,
-        reset_error: str | None = None,
-    ) -> list[AnyComponent]:
-        return self.settings_pages.build_settings_edit_page_components(setting_path, reset_error=reset_error)
-
-    def _build_settings_edit_form(self, setting: EditableSettingDisplay) -> AnyComponent:
-        return self.settings_pages.build_settings_edit_form(setting)
-
-    def _build_settings_reset_form(self, setting_path: str) -> AnyComponent:
-        return self.settings_pages.build_settings_reset_form(setting_path)
-
-    def _get_settings_displays(self) -> tuple[list[EditableSettingDisplay], str | None]:
-        return self.settings_pages.get_settings_displays()
-
-    def _build_settings_badges(self, setting: EditableSettingDisplay) -> list[AnyComponent]:
-        return self.settings_pages.build_settings_badges(setting)
-
-    def _build_settings_value_summary(self, setting: EditableSettingDisplay) -> AnyComponent:
-        return self.settings_pages.build_settings_value_summary(setting)
-
-    def _build_settings_value_cell(self, label: str, value: str) -> AnyComponent:
-        return self.settings_pages.build_settings_value_cell(label, value)
-
-    def _build_settings_edit_guidance(self, setting: EditableSettingDisplay) -> str:
-        return self.settings_pages.build_settings_edit_guidance(setting)
-
-    def _build_settings_patch(self, setting_path: str, raw_value: str) -> JsonObject:
-        return self.settings_pages.build_settings_patch(setting_path, raw_value)
-
-    def _build_dotted_patch(self, dotted_path: str, value: JsonValue) -> JsonObject:
-        return self.settings_pages.build_dotted_patch(dotted_path, value)
-
-    def _persisted_value_matches(self, dotted_path: str, expected_value: object) -> bool:
-        return self.settings_pages.persisted_value_matches(dotted_path, expected_value)
-
-    def _has_persisted_value(self, dotted_path: str) -> bool:
-        return self.settings_pages.has_persisted_value(dotted_path)
-
-    def _lookup_optional_dotted_path(self, root: JsonObject, dotted_path: str) -> object:
-        return self.settings_pages.lookup_optional_dotted_path(root, dotted_path)
-
-    def _format_settings_display_value(self, setting_path: str, value: object) -> str:
-        return self.settings_pages.format_settings_display_value(setting_path, value)
-
-    def _format_settings_provenance(self, provenance: str) -> str:
-        return self.settings_pages.format_settings_provenance(provenance)
-
-    def _build_form_page_components(self, title: str, form_components: list[AnyComponent]) -> list[AnyComponent]:
-        return self.library_pages.build_form_page_components(title=title, form_components=form_components)
-
-    def _build_current_tag_banner_components(self, current_tag_status: CurrentTagStatus | None) -> list[AnyComponent]:
-        return self.library_pages.build_current_tag_banner_components(current_tag_status)
-
-    def _build_disc_library_components(self, discs: list[DiscTable]) -> list[AnyComponent]:
-        return self.library_pages.build_disc_library_components(discs)
-
-    def _build_disc_library_header(self) -> AnyComponent:
-        return self.library_pages._build_disc_library_header()
-
-    def _build_disc_library_row(self, disc: DiscTable) -> AnyComponent:
-        return self.library_pages._build_disc_library_row(disc)
-
-    def _build_disc_header_cell(self, label: str, class_name: str) -> AnyComponent:
-        return self.library_pages._build_disc_header_cell(label, class_name)
-
-    def _build_disc_value_cell(self, label: str, value: str | None, class_name: str) -> AnyComponent:
-        return self.library_pages._build_disc_value_cell(label, value, class_name)
-
-    def _build_new_disc_form_components(self, prefill_current: bool) -> list[AnyComponent]:
-        return self.library_pages.build_new_disc_form_components(prefill_current)
-
-    def _build_edit_disc_form_components(self, tag_id: str) -> list[AnyComponent]:
-        return self.library_pages.build_edit_disc_form_components(tag_id)
-
-    def _build_delete_disc_form_components(self, tag_id: str) -> list[AnyComponent]:
-        return self.library_pages.build_delete_disc_form_components(tag_id)
-
-    async def _current_tag_banner_event_stream(
-        self,
-        request: Request,
-        poll_interval_seconds: float = 0.5,
-    ) -> AsyncIterator[bytes]:
-        async for payload in self.library_pages.current_tag_banner_event_stream(request, poll_interval_seconds):
-            yield payload
-
-    def _serialize_current_tag_components(self, components: list[AnyComponent]) -> str:
-        return self.library_pages.serialize_current_tag_components(components)
 
     @staticmethod
     def _sonos_field_name_for_error(message: str) -> str:
