+++ tests/jukebox/adapters/inbound/admin/test_ui_controller.py
@@ -1,7 +1,7 @@
 import json
 import sys
 from importlib import util
-from unittest.mock import AsyncMock, MagicMock, create_autospec
+from unittest.mock import MagicMock, create_autospec
 
 import pytest
 
@@ -1059,176 +1059,6 @@ def test_ui_controller_does_not_register_get_reset_setting_route():
 
 
 @pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_disc_library_components_render_empty_and_editable_states():
-    from jukebox.adapters.inbound.admin.ui_controller import DiscTable
-
-    controller = build_controller()
-    empty_components = controller._build_disc_library_components([])
-    populated_components = controller._build_disc_library_components(
-        [
-            DiscTable(
-                tag="tag-123",
-                uri="/music/song.mp3",
-                artist="Artist",
-                album="Album",
-                track="Track",
-                shuffle=True,
-            )
-        ]
-    )
-
-    assert empty_components[0].type == "Paragraph"
-    assert empty_components[0].text == "No disc found"
-    edit_button = next(
-        component
-        for component in walk_components(populated_components)
-        if component.type == "Button" and component.text == "Edit ✏️" and component.on_click is not None
-    )
-    assert edit_button.on_click.url == "/discs/tag-123/edit"
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_current_tag_banner_for_unknown_disc_offers_add_cta():
-    from jukebox.domain.entities import CurrentTagStatus
-
-    controller = build_controller()
-
-    components = controller._build_current_tag_banner_components(
-        CurrentTagStatus(tag_id="tag-123", known_in_library=False)
-    )
-    all_components = list(walk_components(components))
-    heading = next(component for component in all_components if component.type == "Heading")
-    button = next(component for component in all_components if component.type == "Button")
-
-    assert heading.text == "Unknown disc on reader"
-    assert button.text == "Add this disc"
-    assert button.on_click.url == "/discs/new?prefill=current"
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_current_tag_banner_for_known_disc_is_informational_only():
-    from jukebox.domain.entities import CurrentTagStatus
-
-    controller = build_controller()
-
-    components = controller._build_current_tag_banner_components(
-        CurrentTagStatus(tag_id="tag-123", known_in_library=True)
-    )
-    all_components = list(walk_components(components))
-    button = next(component for component in all_components if component.type == "Button")
-
-    assert any(component.type == "Heading" and component.text == "Known disc on reader" for component in all_components)
-    assert button.text == "Edit this disc"
-    assert button.on_click.url == "/discs/tag-123/edit"
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_new_disc_form_components_render_blank_add_form():
-    controller = build_controller()
-
-    components = controller._build_new_disc_form_components(prefill_current=False)
-    form = components[0]
-
-    assert form.type == "ModelForm"
-    assert form.submit_url == "/api/ui/discs"
-    assert form.initial is None
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_new_disc_form_components_can_prefill_current_tag():
-    from jukebox.domain.entities import CurrentTagStatus
-
-    controller = build_controller()
-    controller.get_current_tag_status.execute.return_value = CurrentTagStatus(tag_id="tag-123", known_in_library=False)
-
-    components = controller._build_new_disc_form_components(prefill_current=True)
-    form = components[0]
-
-    assert form.type == "ModelForm"
-    assert form.submit_url == "/api/ui/discs"
-    assert form.initial == {"tag": "tag-123", "shuffle": False}
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_edit_disc_form_components_prefill_existing_disc():
-    from jukebox.domain.entities import Disc, DiscMetadata, DiscOption
-
-    controller = build_controller()
-    controller.get_disc.execute.return_value = Disc(
-        uri="/music/song.mp3",
-        metadata=DiscMetadata(artist="Artist", album="Album", track="Track"),
-        option=DiscOption(shuffle=True),
-    )
-
-    components = controller._build_edit_disc_form_components("tag-123")
-    form = components[0]
-
-    assert form.type == "ModelForm"
-    assert form.submit_url == "/api/ui/discs/tag-123"
-    assert form.initial == {
-        "tag": "tag-123",
-        "uri": "/music/song.mp3",
-        "artist": "Artist",
-        "album": "Album",
-        "track": "Track",
-        "shuffle": True,
-    }
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_disc_form_helpers_return_errors_for_invalid_current_tag_state_or_missing_edit_target():
-    from jukebox.domain.entities import CurrentTagStatus
-
-    controller = build_controller()
-
-    controller.get_current_tag_status.execute.return_value = None
-    no_tag_components = controller._build_new_disc_form_components(prefill_current=True)
-    controller.get_current_tag_status.execute.return_value = CurrentTagStatus(tag_id="tag-123", known_in_library=True)
-    known_tag_components = controller._build_new_disc_form_components(prefill_current=True)
-    missing_tag_components = controller._build_edit_disc_form_components("")
-    controller.get_disc.execute.side_effect = ValueError("Missing disc")
-    missing_disc_components = controller._build_edit_disc_form_components("tag-123")
-
-    assert no_tag_components[0].type == "Error"
-    assert known_tag_components[0].type == "Error"
-    assert missing_tag_components[0].type == "Error"
-    assert missing_disc_components[0].type == "Error"
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_form_page_components_include_back_link_and_form():
-    controller = build_controller()
-    components = controller._build_form_page_components(
-        title="Add disc",
-        form_components=controller._build_new_disc_form_components(prefill_current=False),
-    )
-
-    page_components = list(walk_components(components))
-    assert components[0].type == "Page"
-    assert any(component.type == "Heading" and component.text == "Add disc" for component in page_components)
-    assert any(component.type == "ModelForm" for component in page_components)
-    assert any(component.type == "Link" and component.on_click.url == "/" for component in page_components)
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-@pytest.mark.anyio
-async def test_current_tag_banner_event_stream_emits_serialized_updates():
-    from jukebox.domain.entities import CurrentTagStatus
-
-    controller = build_controller()
-    controller.get_current_tag_status.execute.side_effect = [CurrentTagStatus(tag_id="tag-123", known_in_library=False)]
-    request = MagicMock()
-    request.is_disconnected = AsyncMock(side_effect=[False])
-
-    stream = controller._current_tag_banner_event_stream(request, poll_interval_seconds=0)
-    # Avoid the Python 3.10+ `anext` builtin because this repo still supports Python 3.9.
-    first_chunk = await stream.__anext__()
-
-    assert first_chunk.decode("utf-8").startswith("data: [")
-    assert "Unknown disc on reader" in first_chunk.decode("utf-8")
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
 @pytest.mark.anyio
 async def test_create_disc_returns_success_toast():
     from jukebox.adapters.inbound.admin.ui_controller import DiscForm
@@ -1400,18 +1230,6 @@ async def test_delete_disc_returns_404_when_disc_not_found():
 
     assert err.value.status_code == 404
     assert "Disc not found" in err.value.detail
-
-
-@pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
-def test_index_page_shows_remove_toast():
-    controller = build_controller()
-    components = controller._build_index_page_components(toast="toast-remove-disc-success")
-    all_components = list(walk_components(components))
-
-    remove_toast = next(
-        component for component in all_components if component.type == "Toast" and "removed" in str(component.body)
-    )
-    assert remove_toast.open_trigger.name == "toast-remove-disc-success"
 
 
 @pytest.mark.skipif(not FASTUI_INSTALLED, reason="FastUI dependencies are not installed")
